requirements.txt
pyreadr
pandas
numpy
neo4j
openai
python-dotenv
pyarrow
